python/conda_requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
aniso8601=9.0.1=pyhd8ed1ab_0
attrs=20.3.0=pyhd3deb0d_0
boost-cpp=1.74.0=h54f0996_2
branca=0.4.2=pyhd8ed1ab_0
brotlipy=0.7.0=py39hb82d6ee_1001
bzip2=1.0.8=h8ffe710_4
ca-certificates=2021.5.30=h5b45459_0
cairo=1.16.0=hb19e0ff_1008
certifi=2021.5.30=py39hcbf5309_0
cffi=1.14.5=py39h0878f49_0
cfitsio=3.470=h0af3d06_7
chardet=4.0.0=py39hcbf5309_1
click=7.1.2=pyh9f0ad1d_0
click-plugins=1.1.1=py_0
cligj=0.7.1=pyhd8ed1ab_0
cryptography=3.4.7=py39hd8d06c1_0
curl=7.76.1=hf1763fc_1
cycler=0.10.0=py_2
decorator=5.0.7=pyhd8ed1ab_0
descartes=1.1.0=py_4
expat=2.3.0=h39d44d4_0
fiona=1.8.19=py39h9f1b043_0
flask=1.1.2=pyh9f0ad1d_0
flask-restful=0.3.8=py_0
folium=0.12.0=pyhd8ed1ab_1
fontconfig=2.13.1=h1989441_1005
freetype=2.10.4=h546665d_1
freexl=1.0.6=ha8e266a_0
gdal=3.2.2=py39h6795fcd_3
geographiclib=1.50=py_0
geopandas=0.9.0=pyhd8ed1ab_0
geopy=2.1.0=pyhd3deb0d_0
geos=3.9.1=h39d44d4_2
geotiff=1.6.0=h8e90983_5
gettext=0.19.8.1=h1a89ca6_1005
hdf4=4.2.13=h0e5069d_1005
hdf5=1.10.6=nompi_h5268f04_1114
icu=68.1=h0e60522_0
idna=2.10=pyh9f0ad1d_0
intel-openmp=2021.2.0=h57928b3_616
itsdangerous=1.1.0=py_0
jinja2=2.11.3=pyh44b312d_0
joblib=1.0.1=pyhd8ed1ab_0
jpeg=9d=h8ffe710_0
kealib=1.4.14=h96bfa42_2
kiwisolver=1.3.1=py39h2e07f2f_1
krb5=1.17.2=hbae68bd_0
lcms2=2.12=h2a16943_0
libblas=3.9.0=8_mkl
libcblas=3.9.0=8_mkl
libcurl=7.76.1=hf1763fc_1
libffi=3.3=h0e60522_2
libgdal=3.2.2=hade8cfc_3
libglib=2.68.1=h1e62bf3_0
libiconv=1.16=he774522_0
libkml=1.3.0=h9859afa_1013
liblapack=3.9.0=8_mkl
libnetcdf=4.8.0=nompi_hee75d68_101
libpng=1.6.37=h1d00b33_2
libpq=13.2=h4f54205_2
librttopo=1.1.0=hb340de5_6
libspatialindex=1.9.3=h39d44d4_3
libspatialite=5.0.1=h6b539a6_4
libssh2=1.9.0=h680486a_6
libtiff=4.2.0=hc10be44_1
libwebp-base=1.2.0=h8ffe710_2
libxml2=2.9.10=hf5bbc77_4
libzip=1.7.3=hfed4ece_0
lz4-c=1.9.3=h8ffe710_0
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markupsafe=1.1.1=py39hb82d6ee_3
matplotlib-base=3.4.1=py39h581301d_0
mkl=2020.4=hb70f87d_311
msys2-conda-epoch=20160418=1
munch=2.5.0=py_0
networkx=2.5=py_0
numpy=1.20.2=py39h6635163_0
olefile=0.46=pyh9f0ad1d_1
openjpeg=2.4.0=h48faf41_0
openssl=1.1.1k=h8ffe710_0
osmnx=1.0.1=pyhd3deb0d_0
overpy=0.6=pyhd8ed1ab_0
pandas=1.2.4=py39h2e25243_0
patsy=0.5.1=py_0
pcre=8.44=ha925a31_0
pillow=8.1.2=py39h1a9d4f7_1
pip=21.1.1=pyhd8ed1ab_0
pixman=0.40.0=h8ffe710_0
poppler=21.03.0=h9ff6ed8_0
poppler-data=0.4.10=0
postgresql=13.2=h0f1a9bc_2
proj=8.0.0=h1cfcee9_0
pycparser=2.20=pyh9f0ad1d_2
pyopenssl=20.0.1=pyhd8ed1ab_0
pyparsing=2.4.7=pyh9f0ad1d_0
pyproj=3.0.1=py39h1007a03_1
pysocks=1.7.1=py39hcbf5309_3
python=3.9.2=h7840368_0_cpython
python-dateutil=2.8.1=py_0
python_abi=3.9=1_cp39
pytz=2021.1=pyhd8ed1ab_0
requests=2.25.1=pyhd3deb0d_0
rtree=0.9.7=py39h09fdee3_1
scikit-learn=0.24.2=py39he931e04_0
scipy=1.6.3=py39hc0c34ad_0
seaborn=0.11.1=hd8ed1ab_1
seaborn-base=0.11.1=pyhd8ed1ab_1
setuptools=49.6.0=py39hcbf5309_3
shapely=1.7.1=py39h90c6b7e_4
six=1.15.0=pyh9f0ad1d_0
sqlite=3.35.5=h8ffe710_0
statsmodels=0.12.2=py39h5d4886f_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
tiledb=2.2.8=hf84e3da_0
tk=8.6.10=h8ffe710_1
tornado=6.1=py39hb82d6ee_1
tqdm=4.61.1=pyhd8ed1ab_0
tzdata=2021a=he74cb21_0
urllib3=1.26.4=pyhd8ed1ab_0
vc=14.2=hb210afc_4
vs2015_runtime=14.28.29325=h5e1d092_4
werkzeug=1.0.1=pyh9f0ad1d_0
wheel=0.36.2=pyhd3deb0d_0
win_inet_pton=1.1.0=py39hcbf5309_2
wincertstore=0.2=py39hcbf5309_1006
xerces-c=3.2.3=h0e60522_2
xz=5.2.5=h62dcd97_1
zlib=1.2.11=h62dcd97_1010
zstd=1.4.9=h6255e5f_0
